fix compile win64

+++ contrib/Netgen/libsrc/general/dynamicmem.cpp
@@ -148,7 +148,7 @@ namespace netgen
 
     while (p)
       {
-        long unsigned hptr = (long unsigned) (p->ptr);
+        long unsigned hptr = (long long unsigned) (p->ptr);
 	// uintptr_t hptr = reinterpret_cast<uintptr_t>(p->ptr); //??
 
 	hptr /= (1024*1024);
